gateway-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
httpx==0.23.3
python-jose==3.3.0
python-multipart==0.0.6
pydantic==1.10.7
python-dotenv==1.0.0
